requirements.txt
streamlit
langchain
langchain-community
langchain-core
python-dotenv
langchain_groq
scholarly
autogen
transformers
requests
xmltodict
asyncio
